Work towards analytical pressure
Add a 1/x2 method and a integrating polynomial method
to peice_poly. New approach to gravity function. Attempt
at new pressure function.

The general idea is to create a function that is integratable
to give g(r) and then use that to get P(r). The current problem
is that g(r) has a 1/r^2 term outside the integral and this puts
r^-2 terms in the function - and we currently do not handle negative
powers.